STY: extracted DMSP IVM methods
Extracted methods from DMSP into methods file.
